+++ headers/SimplePoly.h
@@ -26,4 +26,4 @@ int get1NumByDegree(Polyonym * poly, int targetDeg);
 
 char get1var(Polyonym * poly);
 
-void change1polyonym(Polyonym * poly,int grade,int in);
+void change1polyonym(Polyonym * poly,int grade,double in);
